파일 정리 및 tools.jar 삭제
