+++ MySmallBasic/run.bat
@@ -1,4 +1,6 @@
-@echo off
-java -cp "./lib/*;./bin/" com.coducation.smallbasic.VersionInfo
-java -cp "./lib/*;./bin/" com.coducation.smallbasic.gui.MySmallBasicGUI
+#!/bin/bash
 
+java -cp "./lib/*:./bin/" com.coducation.smallbasic.VersionInfo
+java -cp "./lib/*:./bin/" com.coducation.smallbasic.gui.MySmallBasicGUI
+cd sbparser
+./sbparser-exe.exe
