+++ MySmallBasic/run.sh
@@ -2,4 +2,5 @@
 
 java -cp "./lib/*:./bin/" com.coducation.smallbasic.VersionInfo
 java -cp "./lib/*:./bin/" com.coducation.smallbasic.gui.MySmallBasicGUI
-
+cd sbparser
+./sbparser-exe.exe
