+++ MySmallBasic/sbparser/goto_table.txt
@@ -1,0 +1,222 @@
+0	Prog	1
+0	MoreThanOneStmt	2
+0	Stmt	3
+0	ExprStatement	4
+5	Expr	12
+5	CondExpr	13
+5	OrExpr	14
+5	AndExpr	15
+5	EqNeqExpr	16
+5	CompExpr	17
+5	AdditiveExpr	18
+5	MultiplicativeExpr	19
+5	UnaryExpr	20
+5	Primary	21
+6	Idxs	27
+10	Expr	36
+10	CondExpr	13
+10	OrExpr	14
+10	AndExpr	15
+10	EqNeqExpr	16
+10	CompExpr	17
+10	AdditiveExpr	18
+10	MultiplicativeExpr	19
+10	UnaryExpr	20
+10	Primary	21
+11	MoreThanOneStmt	37
+11	Stmt	3
+11	ExprStatement	4
+12	CRStmtCRs	38
+22	Idxs	52
+23	Expr	54
+23	CondExpr	13
+23	OrExpr	14
+23	AndExpr	15
+23	EqNeqExpr	16
+23	CompExpr	17
+23	AdditiveExpr	18
+23	MultiplicativeExpr	19
+23	UnaryExpr	20
+23	Primary	21
+24	Primary	55
+29	Expr	57
+29	CondExpr	13
+29	OrExpr	14
+29	AndExpr	15
+29	EqNeqExpr	16
+29	CompExpr	17
+29	AdditiveExpr	18
+29	MultiplicativeExpr	19
+29	UnaryExpr	20
+29	Primary	21
+32	Expr	60
+32	CondExpr	13
+32	OrExpr	14
+32	AndExpr	15
+32	EqNeqExpr	16
+32	CompExpr	17
+32	AdditiveExpr	18
+32	MultiplicativeExpr	19
+32	UnaryExpr	20
+32	Primary	21
+35	CRStmtCRs	62
+39	TheRest	66
+39	Stmt	65
+39	ExprStatement	4
+40	AndExpr	67
+40	EqNeqExpr	16
+40	CompExpr	17
+40	AdditiveExpr	18
+40	MultiplicativeExpr	19
+40	UnaryExpr	20
+40	Primary	21
+41	EqNeqExpr	68
+41	CompExpr	17
+41	AdditiveExpr	18
+41	MultiplicativeExpr	19
+41	UnaryExpr	20
+41	Primary	21
+42	CompExpr	69
+42	AdditiveExpr	18
+42	MultiplicativeExpr	19
+42	UnaryExpr	20
+42	Primary	21
+43	CompExpr	70
+43	AdditiveExpr	18
+43	MultiplicativeExpr	19
+43	UnaryExpr	20
+43	Primary	21
+44	AdditiveExpr	71
+44	MultiplicativeExpr	19
+44	UnaryExpr	20
+44	Primary	21
+45	AdditiveExpr	72
+45	MultiplicativeExpr	19
+45	UnaryExpr	20
+45	Primary	21
+46	AdditiveExpr	73
+46	MultiplicativeExpr	19
+46	UnaryExpr	20
+46	Primary	21
+47	AdditiveExpr	74
+47	MultiplicativeExpr	19
+47	UnaryExpr	20
+47	Primary	21
+48	MultiplicativeExpr	75
+48	UnaryExpr	20
+48	Primary	21
+49	MultiplicativeExpr	76
+49	UnaryExpr	20
+49	Primary	21
+50	UnaryExpr	77
+50	Primary	21
+51	UnaryExpr	78
+51	Primary	21
+56	Expr	81
+56	CondExpr	13
+56	OrExpr	14
+56	AndExpr	15
+56	EqNeqExpr	16
+56	CompExpr	17
+56	AdditiveExpr	18
+56	MultiplicativeExpr	19
+56	UnaryExpr	20
+56	Primary	21
+61	Expr	85
+61	CondExpr	13
+61	OrExpr	14
+61	AndExpr	15
+61	EqNeqExpr	16
+61	CompExpr	17
+61	AdditiveExpr	18
+61	MultiplicativeExpr	19
+61	UnaryExpr	20
+61	Primary	21
+63	CRStmtCRs	87
+82	Expr	90
+82	CondExpr	13
+82	OrExpr	14
+82	AndExpr	15
+82	EqNeqExpr	16
+82	CompExpr	17
+82	AdditiveExpr	18
+82	MultiplicativeExpr	19
+82	UnaryExpr	20
+82	Primary	21
+83	Exprs	92
+83	MoreThanOneExpr	93
+83	Expr	91
+83	CondExpr	13
+83	OrExpr	14
+83	AndExpr	15
+83	EqNeqExpr	16
+83	CompExpr	17
+83	AdditiveExpr	18
+83	MultiplicativeExpr	19
+83	UnaryExpr	20
+83	Primary	21
+84	Idxs	94
+87	MoreThanZeroElseIf	96
+87	OptionalElse	97
+88	TheRest	101
+88	Stmt	65
+88	ExprStatement	4
+89	Exprs	102
+89	MoreThanOneExpr	93
+89	Expr	91
+89	CondExpr	13
+89	OrExpr	14
+89	AndExpr	15
+89	EqNeqExpr	16
+89	CompExpr	17
+89	AdditiveExpr	18
+89	MultiplicativeExpr	19
+89	UnaryExpr	20
+89	Primary	21
+95	Expr	105
+95	CondExpr	13
+95	OrExpr	14
+95	AndExpr	15
+95	EqNeqExpr	16
+95	CompExpr	17
+95	AdditiveExpr	18
+95	MultiplicativeExpr	19
+95	UnaryExpr	20
+95	Primary	21
+98	Expr	106
+98	CondExpr	13
+98	OrExpr	14
+98	AndExpr	15
+98	EqNeqExpr	16
+98	CompExpr	17
+98	AdditiveExpr	18
+98	MultiplicativeExpr	19
+98	UnaryExpr	20
+98	Primary	21
+100	CRStmtCRs	107
+103	MoreThanOneExpr	109
+103	Expr	91
+103	CondExpr	13
+103	OrExpr	14
+103	AndExpr	15
+103	EqNeqExpr	16
+103	CompExpr	17
+103	AdditiveExpr	18
+103	MultiplicativeExpr	19
+103	UnaryExpr	20
+103	Primary	21
+105	OptStep	110
+110	CRStmtCRs	114
+111	Expr	115
+111	CondExpr	13
+111	OrExpr	14
+111	AndExpr	15
+111	EqNeqExpr	16
+111	CompExpr	17
+111	AdditiveExpr	18
+111	MultiplicativeExpr	19
+111	UnaryExpr	20
+111	Primary	21
+112	CRStmtCRs	116
+116	MoreThanZeroElseIf	118
+116	OptionalElse	97
