+++ MySmallBasic/sbparser/mygrammar.grm
@@ -1,0 +1,67 @@
+(CFG "Prog'" [
+ ProductionRule "Prog'" [Nonterminal "Prog"],
+ ProductionRule "Prog" [Nonterminal "MoreThanOneStmt"],
+ ProductionRule "Stmt" [Nonterminal "ExprStatement"],
+ ProductionRule "Stmt" [Terminal "While", Nonterminal "Expr", Nonterminal "CRStmtCRs", Terminal "EndWhile"],
+ ProductionRule "Stmt" [Terminal "ID", Terminal ":"],
+ ProductionRule "Stmt" [Terminal "Goto", Terminal "ID"],
+ ProductionRule "Stmt" [Terminal "For", Terminal "ID", Terminal "=", Nonterminal "Expr", Terminal "To", Nonterminal "Expr", Nonterminal "OptStep", Nonterminal "CRStmtCRs", Terminal "EndFor"],
+ ProductionRule "Stmt" [Terminal "Sub", Terminal "ID", Nonterminal "CRStmtCRs", Terminal "EndSub"],
+ ProductionRule "Stmt" [Terminal "If", Nonterminal "Expr", Terminal "Then", Nonterminal "CRStmtCRs", Nonterminal "MoreThanZeroElseIf"],
+ ProductionRule "Stmt" [],
+ ProductionRule "MoreThanZeroElseIf" [Nonterminal "OptionalElse"],
+ ProductionRule "MoreThanZeroElseIf" [Terminal "ElseIf", Nonterminal "Expr", Terminal "Then", Nonterminal "CRStmtCRs", Nonterminal "MoreThanZeroElseIf"],
+ ProductionRule "OptionalElse" [Terminal "EndIf"],
+ ProductionRule "OptionalElse" [Terminal "Else", Nonterminal "CRStmtCRs", Terminal "EndIf"],
+ ProductionRule "ExprStatement" [Terminal "ID", Terminal "=", Nonterminal "Expr"],
+ ProductionRule "ExprStatement" [Terminal "ID", Terminal ".", Terminal "ID", Terminal "=", Nonterminal "Expr"],
+ ProductionRule "ExprStatement" [Terminal "ID", Terminal ".", Terminal "ID", Terminal "(", Nonterminal "Exprs", Terminal ")"],
+ ProductionRule "ExprStatement" [Terminal "ID", Terminal "(", Terminal ")"],
+ ProductionRule "ExprStatement" [Terminal "ID", Nonterminal "Idxs", Terminal "=", Nonterminal "Expr"],
+ ProductionRule "CRStmtCRs" [Terminal "CR", Nonterminal "TheRest"],
+ ProductionRule "TheRest" [],
+ ProductionRule "TheRest" [Nonterminal "Stmt", Terminal "CR", Nonterminal "TheRest"],
+ ProductionRule "MoreThanOneStmt" [Nonterminal "Stmt"],
+ ProductionRule "MoreThanOneStmt" [Nonterminal "Stmt", Terminal "CR", Nonterminal "MoreThanOneStmt"],
+ ProductionRule "OptStep" [],
+ ProductionRule "OptStep" [Terminal "Step", Nonterminal "Expr"],
+ ProductionRule "Expr" [Nonterminal "CondExpr"],
+ ProductionRule "Exprs" [],
+ ProductionRule "Exprs" [Nonterminal "MoreThanOneExpr"],
+ ProductionRule "MoreThanOneExpr" [Nonterminal "Expr"],
+ ProductionRule "MoreThanOneExpr" [Nonterminal "Expr", Terminal ",", Nonterminal "MoreThanOneExpr"],
+ ProductionRule "CondExpr" [Nonterminal "OrExpr"],
+ ProductionRule "OrExpr" [Nonterminal "OrExpr", Terminal "Or", Nonterminal "AndExpr"],
+ ProductionRule "OrExpr" [Nonterminal "AndExpr"],
+ ProductionRule "AndExpr" [Nonterminal "AndExpr", Terminal "And", Nonterminal "EqNeqExpr"],
+ ProductionRule "AndExpr" [Nonterminal "EqNeqExpr"],
+ ProductionRule "EqNeqExpr" [Nonterminal "EqNeqExpr", Terminal "=", Nonterminal "CompExpr"],
+ ProductionRule "EqNeqExpr" [Nonterminal "EqNeqExpr", Terminal "<>", Nonterminal "CompExpr"],
+ ProductionRule "EqNeqExpr" [Nonterminal "CompExpr"],
+ ProductionRule "CompExpr" [Nonterminal "CompExpr", Terminal "<", Nonterminal "AdditiveExpr"],
+ ProductionRule "CompExpr" [Nonterminal "CompExpr", Terminal "<=", Nonterminal "AdditiveExpr"],
+ ProductionRule "CompExpr" [Nonterminal "CompExpr", Terminal ">", Nonterminal "AdditiveExpr"],
+ ProductionRule "CompExpr" [Nonterminal "CompExpr", Terminal ">=", Nonterminal "AdditiveExpr"],
+ ProductionRule "CompExpr" [Nonterminal "AdditiveExpr"],
+ ProductionRule "AdditiveExpr" [Nonterminal "AdditiveExpr", Terminal "+", Nonterminal "MultiplicativeExpr"],
+ ProductionRule "AdditiveExpr" [Nonterminal "AdditiveExpr", Terminal "-", Nonterminal "MultiplicativeExpr"],
+ ProductionRule "AdditiveExpr" [Nonterminal "MultiplicativeExpr"],
+ ProductionRule "MultiplicativeExpr" [Nonterminal "MultiplicativeExpr", Terminal "*", Nonterminal "UnaryExpr"],
+ ProductionRule "MultiplicativeExpr" [Nonterminal "MultiplicativeExpr", Terminal "/", Nonterminal "UnaryExpr"],
+ ProductionRule "MultiplicativeExpr" [Nonterminal "UnaryExpr"],
+ ProductionRule "UnaryExpr" [Terminal "-", Nonterminal "Primary"],
+ ProductionRule "UnaryExpr" [Nonterminal "Primary"],
+ ProductionRule "Primary" [Terminal "NUM"],
+ ProductionRule "Primary" [Terminal "STR"],
+ ProductionRule "Primary" [Terminal "(", Nonterminal "Expr", Terminal ")"],
+ ProductionRule "Primary" [Terminal "ID"],
+ ProductionRule "Primary" [Terminal "ID", Terminal ".", Terminal "ID"],
+ ProductionRule "Primary" [Terminal "ID", Terminal ".", Terminal "ID", Terminal "(", Nonterminal "Exprs", Terminal ")"],
+ ProductionRule "Primary" [Terminal "ID", Nonterminal "Idxs"],
+ ProductionRule "Idxs" [Terminal "[", Nonterminal "Expr", Terminal "]"],
+ ProductionRule "Idxs" [Terminal "[", Nonterminal "Expr", Terminal "]", Nonterminal "Idxs"]
+ ]
+,
+TokenAttrs [],
+ProdRuleAttrs [],
+"$")
