+++ MySmallBasic/sbparser/prod_rules.txt
@@ -1,0 +1,61 @@
+0: Prog' -> Prog
+1: Prog -> MoreThanOneStmt
+2: Stmt -> ExprStatement
+3: Stmt -> While Expr CRStmtCRs EndWhile
+4: Stmt -> ID :
+5: Stmt -> Goto ID
+6: Stmt -> For ID = Expr To Expr OptStep CRStmtCRs EndFor
+7: Stmt -> Sub ID CRStmtCRs EndSub
+8: Stmt -> If Expr Then CRStmtCRs MoreThanZeroElseIf
+9: Stmt -> 
+10: MoreThanZeroElseIf -> OptionalElse
+11: MoreThanZeroElseIf -> ElseIf Expr Then CRStmtCRs MoreThanZeroElseIf
+12: OptionalElse -> EndIf
+13: OptionalElse -> Else CRStmtCRs EndIf
+14: ExprStatement -> ID = Expr
+15: ExprStatement -> ID . ID = Expr
+16: ExprStatement -> ID . ID ( Exprs )
+17: ExprStatement -> ID ( )
+18: ExprStatement -> ID Idxs = Expr
+19: CRStmtCRs -> CR TheRest
+20: TheRest -> 
+21: TheRest -> Stmt CR TheRest
+22: MoreThanOneStmt -> Stmt
+23: MoreThanOneStmt -> Stmt CR MoreThanOneStmt
+24: OptStep -> 
+25: OptStep -> Step Expr
+26: Expr -> CondExpr
+27: Exprs -> 
+28: Exprs -> MoreThanOneExpr
+29: MoreThanOneExpr -> Expr
+30: MoreThanOneExpr -> Expr , MoreThanOneExpr
+31: CondExpr -> OrExpr
+32: OrExpr -> OrExpr Or AndExpr
+33: OrExpr -> AndExpr
+34: AndExpr -> AndExpr And EqNeqExpr
+35: AndExpr -> EqNeqExpr
+36: EqNeqExpr -> EqNeqExpr = CompExpr
+37: EqNeqExpr -> EqNeqExpr <> CompExpr
+38: EqNeqExpr -> CompExpr
+39: CompExpr -> CompExpr < AdditiveExpr
+40: CompExpr -> CompExpr <= AdditiveExpr
+41: CompExpr -> CompExpr > AdditiveExpr
+42: CompExpr -> CompExpr >= AdditiveExpr
+43: CompExpr -> AdditiveExpr
+44: AdditiveExpr -> AdditiveExpr + MultiplicativeExpr
+45: AdditiveExpr -> AdditiveExpr - MultiplicativeExpr
+46: AdditiveExpr -> MultiplicativeExpr
+47: MultiplicativeExpr -> MultiplicativeExpr * UnaryExpr
+48: MultiplicativeExpr -> MultiplicativeExpr / UnaryExpr
+49: MultiplicativeExpr -> UnaryExpr
+50: UnaryExpr -> - Primary
+51: UnaryExpr -> Primary
+52: Primary -> NUM
+53: Primary -> STR
+54: Primary -> ( Expr )
+55: Primary -> ID
+56: Primary -> ID . ID
+57: Primary -> ID . ID ( Exprs )
+58: Primary -> ID Idxs
+59: Idxs -> [ Expr ]
+60: Idxs -> [ Expr ] Idxs
