+++ MySmallBasic/sbparser/yapb.config
@@ -1,0 +1,11 @@
+Configuration {
+config_SIMPLE = True,
+config_R_LEVEL = 1,
+config_GS_LEVEL = 9,
+config_DEBUG = False,
+config_DISPLAY = False,
+config_PRESENTATION = 0,
+config_ALGORITHM = 3,
+config_COLLECT = False,
+config_TABSTATE = True
+}
