+++ MySmallBasic/src/com/coducation/smallbasic/yapbconfig/YapbConfiguration.java
@@ -1,0 +1,9 @@
+package com.coducation.smallbasic.yapbconfig;
+
+import java.util.HashMap;
+
+public class YapbConfiguration {
+	public String title;
+	public HashMap<String,String> keyValueMap;
+}
+
