--- MySmallBasic/src/com/conducation/smallbasic/yapbconfig/YapbConfiguration.java
@@ -1,9 +1,0 @@
-package com.conducation.smallbasic.yapbconfig;
-
-import java.util.HashMap;
-
-public class YapbConfiguration {
-	public String title;
-	public HashMap<String,String> keyValueMap;
-}
-
